toCows/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
mysqlclient==2.1.1
postgres==4.0
psycopg2-binary==2.9.5
psycopg2-pool==1.1
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
